requirements.txt
watchdog
sentence-transformers
PyMuPDF
python-docx
numpy
